created sql statements for db and table creation (w/ sample data)
Tables that I created here that are not assigned to me are for testing and development purposes only. They can be replaced by the final tables later on. Data in the insert statements are also generated through GEMINI, feeding the create statements into it and allowing it to create 10 entries each. Keep this in mind when using the test data.